factor out buffer.c to its own internal lib
Needed to benchmark C code.
