+++ src/redisearch_rs/c_entrypoint/varint_ffi/cbindgen.toml
@@ -2,7 +2,7 @@ language = "C"
 autogen_warning = "/* Warning, this file is autogenerated by cbindgen from `src/redisearch_rs/c_entrypoint/varint/build.rs. Don't modify it manually. */"
 cpp_compat = true
 pragma_once = true
-includes = ["buffer.h", "redisearch.h"]
+includes = ["buffer/buffer.h", "redisearch.h"]
 
 [parse]
 parse_deps = true
