+++ tests/ctests/test_cntokenize.c
@@ -10,7 +10,7 @@
 #include "test_util.h"
 #include "time_sample.h"
 
-#include "src/buffer.h"
+#include "src/buffer/buffer.h"
 #include "src/index.h"
 #include "src/inverted_index/inverted_index.h"
 #include "src/index_result.h"
