+++ tests/ctests/test_quantile.c
@@ -8,7 +8,7 @@
 */
 
 #include "src/util/quantile.h"
-#include "src/buffer.h"
+#include "src/buffer/buffer.h"
 #include "rmutil/alloc.h"
 #include "test_util.h"
